+++ color.go
@@ -1,5 +1,7 @@
 package colt
 
+import "image/color"
+
 // https://en.wikipedia.org/wiki/Relative_luminance
 const luminanceRed float32 = 0.2126
 const luminanceGreen float32 = 0.7152
@@ -16,6 +18,46 @@ type RGB [3]float32
 
 // RGBA represents a linear RGB color with alpha.
 type RGBA [4]float32
+
+// Std returns c as a color.RGBA, assuming it is alpha-premultiplied.
+func (c SRGB) Std() color.RGBA {
+	return color.RGBA{c[0], c[1], c[2], 1}
+}
+
+// StdN returns c as a color.NRGBA, assuming it is not alpha-premultiplied.
+func (c SRGB) StdN() color.NRGBA {
+	return color.NRGBA{c[0], c[1], c[2], 1}
+}
+
+// Std returns c as a color.RGBA, assuming it is alpha-premultiplied.
+func (c SRGBA) Std() color.RGBA {
+	return color.RGBA{c[0], c[1], c[2], c[3]}
+}
+
+// StdN returns c as a color.NRGBA, assuming it is not alpha-premultiplied.
+func (c SRGBA) StdN() color.NRGBA {
+	return color.NRGBA{c[0], c[1], c[2], c[3]}
+}
+
+// Std lossily returns c as a color.RGBA64, assuming it is alpha-premultiplied.
+func (c RGB) Std() color.RGBA64 {
+	return color.RGBA64{Standard16(c[0]), Standard16(c[1]), Standard16(c[2]), 1}
+}
+
+// StdN lossily returns c as a color.NRGBA64, assuming it is not alpha-premultiplied.
+func (c RGB) StdN() color.NRGBA64 {
+	return color.NRGBA64{Standard16(c[0]), Standard16(c[1]), Standard16(c[2]), 1}
+}
+
+// Std lossily returns c as a color.RGBA64, assuming it is alpha-premultiplied.
+func (c RGBA) Std() color.RGBA64 {
+	return color.RGBA64{Standard16(c[0]), Standard16(c[1]), Standard16(c[2]), uint16(c[3]*0xffff + 0.5)}
+}
+
+// StdN lossily returns c as a color.NRGBA64, assuming it is not alpha-premultiplied.
+func (c RGBA) StdN() color.NRGBA64 {
+	return color.NRGBA64{Standard16(c[0]), Standard16(c[1]), Standard16(c[2]), uint16(c[3]*0xffff + 0.5)}
+}
 
 // Linear returns the linear RGB representation of this sRGB-encoded color.
 func (c SRGB) Linear() RGB {