+++ CHANGES.md
@@ -1,3 +1,5 @@
 # 1.1.0 ()
 * Added `RGB.Sub()` and `RGBA.Sub()`.
 * Added `SRGB.Alpha()`, `SRGBA.SRGB()` `RGB.Alpha()` and `RGBA.RGB()`.
+* Added `Standard16()` and `Linear16()`, `Linear16AllNew()` and `Standard16AllNew()`.
+* Added `SRGB.Std()`, `SRGB.StdN()`, `SRGBA.Std()`, `SRGBA.StdN()`, `RGB.Std()`, `RGB.StdN()`, `RGBA.Std()` and `RGBA.StdN()`.